requirements.txt
pywikibot==8.0.0
Mastodon.py==1.8.1
sseclient==0.0.27